Part two of sounds removal, Instagib folder rework
The instagib folder is the number one offender for having not only the most incomprehensible sounds in another language, but also sounds that trigger easily and overstay their welcome. This remedies most of the issues between annoying sounds, in different languages, plus some renaming for some more underused sounds to be more accessible.