Finally fix instagib (Probably)
Throw a band-aid on the issue of intrusive sounds by renaming them to something more esoteric, mostly by putting "song" on them, as an alternative to deleting everything. Also special thanks to Zeni for transcription for some of these.

This folder still has a million different problems but I'm not gonna be the one to fix them, hell no