hitting the griddy for ukraine (#404)
* Add files via upload

pussy cats. 250

* new malkavian sounds

now (hopefully) in the right refresh rate!!!

* third attempt at new malkavian sounds

if these end up being the wrong hz me and carlos are going to put ourselves in body bags

* Delete assasinate a man and get somebody killed ill load you up.ogg

typo

* folder creation, variable sounds

first commit for op4

* essential op4 sounds, part 1

i can only upload 100 files at a time :skull:

* essential op4 sounds, part 2

see last commit

* op4 sounds - MASSIVE COMMIT

it was a crime for these sounds not to be present in chatsounds. I expect this commit to fail in several places.

* delete op4 folder

because its easier for me to just do this on desktop :)

* lets try this again. creation of op4 folder

variables go first

* op4 sounds part 2

13 files left

* op4 sounds part 3

the final 13

* we dont need these anymore :)

unused

* stop being ratchet!!!

funny orange cat

* hitting the griddy for ukraine

funne